requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
sentence-transformers>=2.6.0
scikit-learn>=1.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0
